Create courses_gateway app successfully
