+++ Backend/courses_gateway/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoursesGatewayConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'courses_gateway'
